InducingPoints

InducingPoints.jl aims at providing an easy way to select inducing points locations for Sparse Gaussian Processes both in an online and offline setting.

The point selection is splitted in the online (OnIPSA) and offline settings.

All algorithms inherit from AbstractInducingPointsSelection or AIPSA which can be passed to the different APIs

Offline Inducing Points Selection

alg = KMeansAlg(10)
Z = inducingpoints(alg, X; kwargs...)

The Offline options are:

KmeansAlg : use the k-means algorithm to select centroids minimizing the square distance with the dataset. The seeding is done via k-means++. Note that the inducing points are not going to be a subset of the data

kDPP : sample from a k-Determinantal Point Process to select k points. Z will be a subset of X

StdDPP : sample from a standard Determinantal Point Process. The number of inducing points is not fixed here. Z will be a subset of X

RandomSubset : sample randomly k points from the data set uniformly.

Greedy : Will select a subset of X which maximizes the ELBO (in a stochastic way)

Online Inducing Points Selection

Online selection is a bit more involved.

alg = OIPS()
Z = initZ(alg, x_1; kwargs...)
for x in eachbatch(X)
    updateZ!(Z, alg, x; kwargs...)
end

With init, a first instance of Z is created. update! will then update the vectors in place.

The Online options are:

OIPS : A method based on distance between inducing points and data

UniformGrid : A regularly-spaced grid whom edges are adapted given the data

SeqDPP : Sequential Determinantal Point Processes, subsets are regularly sampled from the new data batches conditionned on the existing inducing points.

StreamKmeans : An online version of k-means.

Webscale : Another online version of k-means

Index

InducingPoints.Greedy

InducingPoints.KmeansAlg

InducingPoints.OnlineIPSelection

InducingPoints.RandomSubset

InducingPoints.SeqDPP

InducingPoints.StdDPP

InducingPoints.StreamKmeans

InducingPoints.UniGrid

InducingPoints.Webscale

InducingPoints.kDPP

InducingPoints.distance

InducingPoints.find_nearest_center

InducingPoints.inducingpoints

InducingPoints.initZ

InducingPoints.kmeans_seeding

InducingPoints.updateZ!

API

InducingPoints.Greedy — Type

GreedyIP(m::Int, s::Int)

m is the desired number of inducing points

s is the minibatch size on which to select a new inducing point

Greedy approach first proposed by Titsias[1]. Algorithm loops over minibatches of data and select the best ELBO improvement.

[1] Titsias, M. Variational Learning of Inducing Variables in Sparse Gaussian Processes. Aistats 5, 567–574 (2009).

InducingPoints.KmeansAlg — Type

KmeansAlg(m::Int, metric::SemiMetric=SqEuclidean(); nMarkov = 10, tol = 1e-3)

k-Means [1] initialization on the data X taking m inducing points. The seeding is computed via [2], nMarkov gives the number of MCMC steps for the seeding.

Arguments

k::Int : Number of inducing points

metric::SemiMetric : Metric used to compute the distances for the k-means algorithm

Keyword Arguments

nMarkov::Int : Number of random steps for the seeding

tol::Real : Tolerance for the kmeans algorithm

[1] Arthur, D. & Vassilvitskii, S. k-means++: The advantages of careful seeding. in Proceedings of the eighteenth annual ACM-SIAM symposium on Discrete algorithms 1027–1035 (Society for Industrial and Applied Mathematics, 2007). [2] Bachem, O., Lucic, M., Hassani, S. H. & Krause, A. Fast and Provably Good Seedings for k-Means. Advances in Neural Information Processing Systems 29 55–63 (2016) doi:10.1109/tmtt.2005.863818.

InducingPoints.OnlineIPSelection — Type

OIPS(ρ_accept=0.8; η=0.95, kmax=Inf, kmin=10, ρ_remove=Inf)
OIPS(kmax, η=0.98, kmin=10)

Online Inducing Points Selection. Method from the paper include reference here.

InducingPoints.RandomSubset — Type

RandomSubset(m::Int)

Arguments

m::Int: Number of inducing points

Uniform sampling of a subset of m points ofthe data.

InducingPoints.SeqDPP — Type

SeqDPP()

Sequential sampling via DeterminantalPointProcesses

InducingPoints.StdDPP — Type

StdDPP(kernel::Kernel)

Standard DPP (Determinantal Point Process) sampling given kernel. The size of the returned Z is not fixed (but cannot be empty unlike in a classical DPP)

InducingPoints.StreamKmeans — Type

StreamKmeans(k_target::Int)

Online clustering algorithm [1] to select inducing points in a streaming setting. Reference : [1] Liberty, E., Sriharsha, R. & Sviridenko, M. An Algorithm for Online K-Means Clustering. arXiv:1412.5721 cs.

InducingPoints.UniGrid — Type

UniGrid(K::Int)

where m is the number of points on each dimension Adaptive uniform grid based on [1]

[1] Moreno-Muñoz, P., Artés-Rodríguez, A. & Álvarez, M. A. Continual Multi-task Gaussian Processes. (2019).

InducingPoints.Webscale — Type

Webscale(k::Int)

Online k-means algorithm based on [1].

[1] Sculley, D. Web-scale k-means clustering. in Proceedings of the 19th international conference on World wide web - WWW ’10 1177 (ACM Press, 2010). doi:10.1145/1772690.1772862.

InducingPoints.kDPP — Type

kDPP(m::Int, kernel::Kernel)

k-DPP (Determinantal Point Process) will return a subset of X of size m, according to DPP probability

InducingPoints.distance — Method

Compute the distance (kernel if included) between a point and a findnearestcenter

InducingPoints.find_nearest_center — Function

Find the closest center to X among Z, return the index and the distance

InducingPoints.inducingpoints — Function

 inducingpoints([rng::AbstractRNG], alg::OffIPSA, X::AbstractVector; kwargs...)
 inducingpoints([rng::AbstractRNG], alg::OffIPSA, X::AbstractMatrix; obsdim=1, kwargs...)

Select inducing points according to the algorithm alg.

InducingPoints.initZ — Function

 initZ([rng::AbstractRNG], alg::OnIPSA, X::AbstractVector; kwargs...)
 initZ([rng::AbstractRNG], alg::OnIPSA, X::AbstractMatrix; obsdim=1, kwargs...)

Select inducing points according to the algorithm alg and return a Vector of Vector.

InducingPoints.kmeans_seeding — Method

Fast and efficient seeding for KMeans based on `Fast and Provably Good Seeding for k-Means

InducingPoints.updateZ! — Function

updateZ!([rng::AbstractRNG], Z::AbstractVector, alg::OnIPSA, X::AbstractVector; kwargs...)

Update inducing points Z with data X and algorithm alg